+++ lib/internal/streams/pipeline.js
@@ -114,7 +114,7 @@ function makeAsyncIterable(val) {
 
 async function* fromReadable(val) {
   if (!Readable) {
-    Readable = require('_stream_readable');
+    Readable = require('internal/streams/readable');
   }
 
   yield* Readable.prototype[SymbolAsyncIterator].call(val);
