+++ node.gyp
@@ -236,6 +236,11 @@
       'lib/internal/streams/duplexpair.js',
       'lib/internal/streams/from.js',
       'lib/internal/streams/legacy.js',
+      'lib/internal/streams/readable.js',
+      'lib/internal/streams/writable.js',
+      'lib/internal/streams/duplex.js',
+      'lib/internal/streams/passthrough.js',
+      'lib/internal/streams/transform.js',
       'lib/internal/streams/destroy.js',
       'lib/internal/streams/state.js',
       'lib/internal/streams/pipeline.js',
