+++ test/parallel/test-bootstrap-modules.js
@@ -96,11 +96,11 @@ if (!common.isMainThread) {
     'Internal Binding messaging',
     'Internal Binding symbols',
     'Internal Binding worker',
-    'NativeModule _stream_duplex',
-    'NativeModule _stream_passthrough',
-    'NativeModule _stream_readable',
-    'NativeModule _stream_transform',
-    'NativeModule _stream_writable',
+    'NativeModule internal/streams/duplex',
+    'NativeModule internal/streams/passthrough',
+    'NativeModule internal/streams/readable',
+    'NativeModule internal/streams/transform',
+    'NativeModule internal/streams/writable',
     'NativeModule internal/error_serdes',
     'NativeModule internal/process/worker_thread_only',
     'NativeModule internal/streams/buffer_list',
